patch: patch version number into shared object output
Upstream guarentees that the LTS branch, in this case 3.6
will remain stable, but it doesn't guarentee that a future
LTS branch will be API/ABI stable even under the same major
version number. For example a system with mbedtls3.6 and
mbedtls3.8 will not be stable, and they have been histrocially
bad at updating the soversions and development package so links
will need to be versioned so they point to the right package.

They consider their semantic versioning scheme to only apply
to the API portion, but they re-order structs and other data
structures without regard to ABI stability outside of
maintained LTS branches.

Signed-off-by: Bill Roberts <bill.roberts@arm.com>